+++ tapo/src/responses/token_result.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::responses::TapoResponseExt;
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Serialize, Deserialize)]
 pub(crate) struct TokenResult {
     pub token: String,
 }